pyproject.toml
[project]
name = "hpic-committee"
version = "0.1.0"
description = "AI-Augmented Deliberative Committee for Chicago stadium policy evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[project.scripts]
hpic-simulate = "src.simulate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests that call the real API (deselect with '-m \"not integration\"')",
]
